M2-6042: Move Take Now validation logic to hook
