+++ src/features/TakeNow/lib/TakeNowParams.types.ts
@@ -1,0 +1,6 @@
+export type TakeNowParams = {
+  appletId: string;
+  startActivityOrFlow: string;
+  subjectId: string;
+  respondentId: string;
+};
